+++ packages/fury-adapter-oas3-parser/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Adds support for `termsOfService` in 'Info Object'.
 
+- Adds support for `contact` in 'Info Object'.
+
 ## 0.10.1 (2020-01-30)
 
 ### Bug Fixes